remove MigrateServicePreferenceFromLegacy from io-profile
