--- apps/io-profile/MigrateServicePreferenceFromLegacy/function.json
@@ -1,12 +1,0 @@
-{
-  "bindings": [
-    {
-      "queueName": "%MIGRATE_SERVICES_PREFERENCES_PROFILE_QUEUE_NAME%",
-      "connection": "FN_APP_STORAGE_CONNECTION_STRING",
-      "name": "oldAndNewProfiles",
-      "type": "queueTrigger",
-      "direction": "in"
-    }
-  ],
-  "scriptFile": "../dist/MigrateServicePreferenceFromLegacy/index.js"
-}
